+++ YourssuCalculator/Domain/Model/CalculateTpyeModel.swift
@@ -1,0 +1,16 @@
+//
+//  CalculateTpyeModel.swift
+//  YourssuCalculator
+//
+//  Created by 홍준혁 on 2023/04/27.
+//
+
+import Foundation
+
+@frozen
+enum CalculateTpye {
+    case plus
+    case minus
+    case mulitplus
+    case divis
+}
